requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-google-genai
qdrant-client
pandas